core: add supported for delayed regfile writeback
RocketCore allows the instruction to exit but delays the writeback
of the data to regfile if possible. During the delayed time, following
instructions are allowed to exit as well. This model has some conflicts
with the instruction-level co-simulation, which checks every cycle
the regfile and instruction commits.

This commit we use the DelayedWriteback support from DiffTest to
avoid false positive mismatches.